backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator>=2.0.0
google-generativeai>=0.3.0
